check before delete file

+++ src/ExecuteManager.cs
@@ -184,7 +184,8 @@
                             logger.Info($"file {targetFile} was copied");
                             break;
                         case DistributeMode.DELETEALLFIRST:
-                            File.Delete(targetFile);
+                            if (File.Exists(targetFile))
+                                File.Delete(targetFile);
                             Directory.CreateDirectory(targetPath);
                             File.Copy(path, targetFile, false);
                             break;
@@ -341,7 +342,6 @@
                         throw new Exception($"Unknown hub mode {settings.Mode}");
                     }
                 }
-
                 return null;
             }
             catch (Exception ex)
